municipal-budget-viz/etl/requirements.txt
pdfplumber>=0.11.0
camelot-py[cv]>=0.11.0
pandas>=2.0.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
pytest>=8.0.0
